printer_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
brother_ql==0.9.4
segno==1.6.1
pillow==10.2.0
